Implement redis_lock
This actually changes a pretty large chunk of the API. Instead of
hacking into Redis and including a module, which is really nasty, it
accepts an instance of redis as its first argument, as well as the key.

It doesn't do any time-based retries anymore; instead, it only relies on a
number of retries (which defaults to 10) and sleeps 200ms between
attempting to acquire a lock. This simplifies the logic of key and value
generation immensely and talks to redis a lot less during the entire
lifecycle of the lock.

This also tests different cases a lot more thoroughly since the tests
were written before any of the redis_lock code.